Fix: Empleado -> Telefono (not big enough)

+++ database/migrations/2023_09_27_170814_create_empleados_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->integer("dni")->unique();
             $table->string("email");
             $table->string("direccion");
-            $table->integer("telefono");
+            $table->bigInteger("telefono");
             $table->date("fecha_ingreso");
             $table->timestamps();
         });
